backend/requirements.txt
flask-2.2.2
flask-cors-3.0.10
